Fix/admin portal live matches and teams (#12)
* Fixed Inconsient data between database and schema

* Feat: ported apis from preview admin panel

* feat: Add error and success message handling in CreatePostPage

The code changes in this commit add error and success message handling in the CreatePostPage component. The component now displays a red error message if there is an error creating a post, and a green success message if the post is created successfully.

* Added Post Delete in admin portal

* Added Edit and Delete button

* Made API compaitable with app

* feat: Add PUT route for updating a post

This commit adds a new PUT route in the post API to handle updating a post. The route is protected and can only be accessed by authenticated users with admin privileges. The route expects a JSON payload containing the updated post data, and it updates the corresponding post in the database. If the post is not found, an error is thrown. This new route allows for editing existing posts in the application.

* refactor: Update UserFooter component

This commit refactors the UserFooter component to improve its structure and readability. The changes include:
- Reorganizing the HTML structure of the footer using CSS grid layout.
- Adding semantic HTML tags for better accessibility.
- Updating the footer links to use the Link component from Next.js for improved navigation.
- Adding lazy loading for the social media icons to improve page load performance.
- Updating the app install link to use the Link component for consistent navigation.

These changes enhance the UserFooter component and improve the overall user experience.

* refactor: Update UserFooter component

* Add go live and some api

* Feat: Some mange live page

* fix: Penality Checkbox

* fix: Test

* fix: Match Edit only load when successfully fetches data

* Fixed Edit Live State

* feat: Match Update

* Added Styling

* refactor: Improve EditLivePost component structure and functionality

The EditLivePost component has been refactored to enhance its structure and functionality. The following changes have been made:
- Added the `parse` function from the `html-react-parser` library to parse HTML content in the timeline history.
- Added the `TrashIcon` component from the `@heroicons/react` library for deleting timeline entries.
- Created the `MatchDetailsForm` component to handle the form for editing match details.
- Created the `TimelineHistory` component to display the timeline history with the ability to delete entries.
- Updated the rendering of the `EditLivePostForm` component to include the `MatchDetailsForm` and `TimelineHistory` components.

These changes improve the overall user experience and make the EditLivePost component more maintainable and extensible.

* refactor: Enhance EditLivePost component structure and functionality

* refactor: Enhance EditLivePost component structure and functionality

* refactor: Update PUT function in match route to remove unnecessary console.log statement

* fix: Date problem

* refactor: Enhance EditLivePost component structure and functionality

* refactor: Enhance EditLivePost component structure and functionality

* Fix Missing States

* Fixed Few Issues

* Fix UI

* Final UI Fixes