Fix/user front improvements (#14)
* Added new Loading Bars

* Removed Posts Page

* Fix about-us page

* Alumni UI Changes

* feat: Add API routes for retrieving and creating alumni data

This commit adds API routes for retrieving and creating alumni data. The GET route fetches all alumni data from the database and returns it as a JSON response. The POST route creates a new alumni record in the database based on the request payload. Both routes handle errors and return appropriate error messages and status codes.

* Added new alumni api

* Fixed Alumni Section Apis and Edit Form

* Fix: Removed Taboid link

* feat: Add query parameters for session filtering in alumni API

This commit adds query parameters for session filtering in the alumni API. The `startSession` and `endSession` parameters allow users to specify a range of sessions to filter the alumni data. The API now queries the database based on these parameters and returns the filtered results.

* Added form to add image for alumni

* refactor: Update AlumniPage component and AlumniCard component

This commit refactors the AlumniPage component and AlumniCard component.

- In the AlumniPage component:
  - Adds state hooks for alumniData and loading
  - Implements fetchAlumniData function to fetch alumni data from the API
  - Uses useEffect to call fetchAlumniData on component mount
  - Displays a loading spinner while data is being fetched

- In the AlumniCard component:
  - Updates the import statement for the Alumni model
  - Updates the props to match the new Alumni model structure
  - Removes the position field from the component

These changes improve the code structure and functionality of the Alumni page.

* Alumni Section Added