Fix/general UI and performace improvments (#16)
* fix: Replace Next/Image with AritcleImage to handle errors

* refactor: Update Article component to use ArticleImage for consistency

The Article component was updated to use the ArticleImage component for displaying article images. This change ensures consistency across the codebase and improves error handling for image loading.

* refactor: Update image source in user and admin headers

The image source in the user and admin headers was updated to use the webp format instead of the png format. This change improves page load performance by using a more efficient image format.

* refactor: Update image source in user and admin headers

* fix: Alumni api behavioyr changes

* refactor: Update loadRelatedPosts function to filter out current post

The loadRelatedPosts function in the page.tsx file has been updated to filter out the current post from the related posts. This change ensures that the current post is not displayed as a related post.

* Behavious and small ui changes

* chore: Made Alumni header more respomsive

* Few changes

* Added TOS and Privacy Policy